refactor(account): 🎨 Code refactoring
Code refactoring so that its structure reflects architectural decisions.
