+++ src/modules/account/repositories/PostgresUsersRepository.ts
@@ -1,4 +1,4 @@
-import { User } from "../model/User";
+import { User } from "./../model/User"
 import { ICreateUserDTO, IUsersRepository } from "./IUsersRepository";
