+++ src/modules/account/repositories/UsersRepository.ts
@@ -1,4 +1,4 @@
-import { User } from "../model/User";
+import { User } from "./../model/User";
 import { ICreateUserDTO, IUsersRepository } from "./IUsersRepository";
 
 class UsersRepository implements IUsersRepository {
