+++ src/modules/account/services/CreateUserService.ts
@@ -1,4 +1,4 @@
-import { IUsersRepository } from "../../repositories/IUsersRepository";
+import { IUsersRepository } from "./../repositories/IUsersRepository";
 
 interface IRequest {
     name: string;
